// src/common/enums/relationship-type.enum.ts
export enum RelationshipType {
  Unknown = "unknown",
  Friend = "friend",
  Partner = "partner",
  FamilyMember = "family_member",
  Colleague = "colleague",
  Client = "client",
  Vendor = "vendor",
  VerifiedBusinessContact = "verified_business_contact",
  InnerCircle = "inner_circle",
  HouseholdService = "household_service",
  SupportAgent = "support_agent",
}